+++ CHANGELOG.md
@@ -12,6 +12,7 @@ NEXT_VERSION_BUMP: PATCH
 
 ### CryptoOnramp
 * [CHANGED] Renamed the EU attestation presentation API from `presentCrsCarfDeclaration()` to `presentUserAttestation()`, `crsCarfDeclarationCallback` to `userAttestationCallback`, and renamed `OnrampCrsCarfDeclarationResult` to `OnrampUserAttestationResult`.
+* [ADDED] A new `AppAttestationUnavailableException` was added when configuring onramp for additional information when an error occurs.
 
 ## 23.10.1 - 2026-06-15